separa os códigos de estilo em arquivos específicos
