+++ src/components/ActivityCard.jsx
@@ -1,64 +1,6 @@
-// ActivityCard.jsx
-
 import styled from 'styled-components';
 import { GameController, Book, FileText } from '@phosphor-icons/react';
-
-// Estilos do Card
-const CardContainer = styled.div`
-  background-color: ${props => props.bgColor || '#ff9800'};
-  border-radius: 8px;
-  padding: 14px;
-  display: flex;
-  flex-direction: column;
-  height: 160px;
-`;
-
-const CardHeader = styled.div`
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 8px;
-`;
-
-const CardIcon = styled.div`
-  color: ${props => props.iconColor || '#000'};
-  font-size: 24px;
-  display: flex;
-  align-items: center;
-`;
-
-const CardCategory = styled.span`
-  font-size: 10px;
-  background-color: rgba(255, 255, 255, 0.3);
-  padding: 2px 6px;
-  border-radius: 4px;
-`;
-
-const CardTitle = styled.h3`
-  font-size: 16px;
-  font-weight: 600;
-  margin-bottom: 6px;
-  color: ${props => props.color || '#000'};
-`;
-
-const CardDescription = styled.p`
-  font-size: 11px;
-  flex-grow: 1;
-  color: ${props => props.color || '#000'};
-`;
-
-const CardButton = styled.button`
-  font-size: 10px;
-  background-color: rgba(0, 0, 0, 0.1);
-  border: none;
-  border-radius: 4px;
-  padding: 6px;
-  text-transform: uppercase;
-  cursor: pointer;
-  color: ${props => props.color || '#000'};
-  font-weight: 600;
-  align-self: flex-start;
-  margin-top: 6px;
-`;
+import {CardContainer, CardHeader, CardIcon, CardCategory, CardTitle, CardDescription, CardButton} from '../styles/ActivityCard';
 
 // Mapeamento de tipo para ícone
 const getIconForType = (type) => {
